requirements.txt
azure-ai-ml
azure-identity
azure-mgmt-resource
azure-mgmt-storage
azure-storage-blob
mldesigner
pandas
numpy
pyod
tenacity